modify parameter type in /API/UGA/Quser/delete

+++ src/test/java/com/questetra/bpms/client/swagger/api/QuserApiTest.java
@@ -65,8 +65,8 @@ public class QuserApiTest {
     @Test
     public void deleteTest() throws ApiException {
         Long id = null;
-        BigDecimal delegateQuserId = null;
-        BigDecimal delegateQgroupId = null;
+        Long delegateQuserId = null;
+        Long delegateQgroupId = null;
         api.delete(id, delegateQuserId, delegateQgroupId);
 
         // TODO: test validations
